requirements.txt
streamlit
openai
pillow
requests
diffusers
transformers
torch
torchvision
torchaudio
streamlit-lottie
